Revert "220"
This reverts commit 05689f26bf4327f8210fbe41dcbd6abd6f27bf4c.

+++ app/src/main/java/co/dito/abako/apijitpack/data/model/request/delivery/DllItemCredit.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class DllItemCredit(
 @SerializedName("idArt") val idArt: Int = 0,
-@SerializedName("cant") val cant: Double = 0.0,
+@SerializedName("cant") val cant: Int = 0,
 @SerializedName("prcs") val prcs: String = ""
 )
