+++ app/src/main/java/co/dito/abako/apijitpack/data/model/request/delivery/SetCreditNoteRequest.kt
@@ -51,7 +51,7 @@ fun SetCreditNoteRequest.toCreditNoteRequest(): CreditNoteRequest {
 fun SetCreditNoteDetailRequest.toDllItemCredit(): DllItemCredit {
     return DllItemCredit(
         idArt = this.idArticle,
-        cant = this.quantity,
+        cant = this.quantity.toInt(),
         prcs = this.prcs
     )
 }
